+++ src/parxy_cli/tui/__init__.py
@@ -1,1 +1,6 @@
 """TUI module for Parxy document processing."""
+
+from parxy_cli.tui.app import ParxyTUI, run_tui
+from parxy_cli.tui.widgets import FileTreeSelector, FilteredDirectoryTree
+
+__all__ = ['ParxyTUI', 'run_tui', 'FileTreeSelector', 'FilteredDirectoryTree']
