+++ src/parxy_cli/tui/widgets/__init__.py
@@ -1,0 +1,20 @@
+"""Reusable widgets for the Parxy TUI."""
+
+from parxy_cli.tui.widgets.file_tree_selector import FileTreeSelector, FilteredDirectoryTree
+from parxy_cli.tui.widgets.parser_selector import ParserSelector
+from parxy_cli.tui.widgets.results_viewer import (
+    ResultsViewer,
+    DiffViewer,
+    SideBySideViewer,
+    ParserResults,
+)
+
+__all__ = [
+    'FileTreeSelector',
+    'FilteredDirectoryTree',
+    'ParserSelector',
+    'ResultsViewer',
+    'DiffViewer',
+    'SideBySideViewer',
+    'ParserResults',
+]
